Time: 72 ms (5.05%), Space: 23.2 MB (40.31%) - LeetHub

+++ 0344-reverse-string/0344-reverse-string.cpp
@@ -1,12 +1,10 @@
 class Solution {
 public:
     void reverseString(vector<char>& s) {
-        
-        size_t n = s.size();
+        long long int n = s.size();
         
 		// reverse string by mirror image
         for( size_t i = 0; i < n/2 ; i ++){
-		
             swap(s[i], s[n-1-i]);            
         }
